+++ Changelog.md
@@ -18,6 +18,7 @@ Sorry for the breaking changes, but they are worth it!
 - Added `sequence/state`, a new sequence endpoint for retrieving information that is much more elaborate and also supports plugins.
 - Added `sequence/edit`, which works similary to `profile/set-value`. Note that this mainly supports fields that expect simple types like strings, numbers etc, and may not work for things like enums or objects (filter, time source, ...). This is an **experimental** feature and may be unreliable or uncomplete.
 - Added `mount/set-park-position`, which sets the current mount position as park position.
+- Added `pause-alignment` and `resume-alignment` to the TPPA WebSocket.
 
 ## 2.1.8.0